+++ src/Voron/Platform/Posix/Posix32BitsMemoryMapPager.cs
@@ -13,6 +13,7 @@ using Voron.Data;
 using Voron.Global;
 using Voron.Impl;
 using Voron.Impl.Paging;
+using Voron.Platform.Win32;
 using Voron.Util.Settings;
 
 namespace Voron.Platform.Posix
@@ -34,7 +35,7 @@ namespace Voron.Platform.Posix
         public override long TotalAllocationSize => _totalAllocationSize;
 
         public Posix32BitsMemoryMapPager(StorageEnvironmentOptions options, VoronPathSetting file, long? initialFileSize = null,
-            bool usePageProtection = false) : base(options, usePageProtection)
+            bool usePageProtection = false) : base(options, canPrefetchAhead: false, usePageProtection: usePageProtection)
         {
             _options = options;
             FileName = file;
@@ -78,12 +79,7 @@ namespace Voron.Platform.Posix
 
             NumberOfAllocatedPages = _totalAllocationSize / Constants.Storage.PageSize;
 
-            SetPagerState(new PagerState(this)
-            {
-                Files = null,
-                MapBase = null,
-                AllocationInfos = new PagerState.AllocationInfo[0]
-            });
+            SetPagerState(new PagerState(this, Options.PrefetchSegmentSize, Options.PrefetchResetThreshold));
         }
 
         private static void ThrowNotSupportedOption(string file)
@@ -498,15 +494,14 @@ namespace Voron.Platform.Posix
             return FileName.FullPath;
         }
 
-
-        public override void TryPrefetchingWholeFile()
-        {
-            // we never want to do this, we'll rely on the OS to do it for us
-        }
-
-        public override void MaybePrefetchMemory(List<long> pagesToPrefetch)
+        protected internal override void PrefetchRanges(Win32MemoryMapNativeMethods.WIN32_MEMORY_RANGE_ENTRY* list, int count)
         {
             // we never want to do this here
         }
+        
+        public override void ReleaseAllocationInfo(byte* baseAddress, long size)
+        {
+            ReleaseAllocationInfoWithoutUnmapping(baseAddress, size);            
+        }        
     }
 }
